Simplify EcoDataVirtualizedList component
- Remove CursorPaginationState class, inline all logic
- Use @key for refresh instead of RefreshDataAsync
- No locks needed in single-threaded Blazor WebAssembly
- Generation tracking handles stale requests
- Rename RefreshAsync to Refresh (sync method)
