+++ src/Common/EcoData.Common.Pagination.Blazor/EcoDataVirtualizedList.razor
@@ -23,7 +23,7 @@
     }
     else
     {
-        <Virtualize @ref="_virtualize"
+        <Virtualize @key="_generation"
                     TItem="TItem"
                     ItemsProvider="LoadItemsAsync"
                     ItemSize="ItemSize"
